chore(warnings): sanitize Chameleon user_data casts and signatures

+++ fw/application/src/app/chameleon/scene/chameleon_scene_menu_card_slot_num_select.c
@@ -13,13 +13,14 @@
 
 #include "mui_icons.h"
 #include "tag_helper.h"
+#include <stdint.h>
 
 void chameleon_scene_menu_card_slot_num_select_on_event(mui_list_view_event_t event, mui_list_view_t *p_list_view,
                                                     mui_list_item_t *p_item) {
     app_chameleon_t *app = p_list_view->user_data;
     switch (p_item->icon) {
     case ICON_DATA: {
-        tag_helper_set_slot_num((uint8_t)p_item->user_data);
+        tag_helper_set_slot_num((uint8_t)(uintptr_t)p_item->user_data);
         mui_scene_dispatcher_previous_scene(app->p_scene_dispatcher);
     } break;
 
@@ -33,9 +34,9 @@ void chameleon_scene_menu_card_slot_num_select_on_enter(void *user_data) {
     app_chameleon_t *app = user_data;
 
     // 只添加 8、20、50 三个选项
-    mui_list_view_add_item_ext(app->p_list_view, ICON_DATA, "8", NULL, (void *)8);
-    mui_list_view_add_item_ext(app->p_list_view, ICON_DATA, "20", NULL, (void *)20);
-    mui_list_view_add_item_ext(app->p_list_view, ICON_DATA, "50", NULL, (void *)50);
+    mui_list_view_add_item_ext(app->p_list_view, ICON_DATA, "8", NULL, (void *)(uintptr_t)8);
+    mui_list_view_add_item_ext(app->p_list_view, ICON_DATA, "20", NULL, (void *)(uintptr_t)20);
+    mui_list_view_add_item_ext(app->p_list_view, ICON_DATA, "50", NULL, (void *)(uintptr_t)50);
 
     mui_list_view_add_item(app->p_list_view, ICON_BACK, _T(MAIN_RETURN), NULL_USER_DATA);
